+++ octoprint_octolapse/data/lib/c/gcode_position_processor.h
@@ -60,7 +60,7 @@ extern "C"
 static bool ParsePositionArgs(PyObject *py_args, gcode_position_args *args);
 static bool ParseStabilizationArgs(PyObject *py_args, stabilization_args* args);
 static bool ParseStabilizationArgs_SmartLayer(PyObject *py_args, smart_layer_args* args);
-static bool ExecuteStabilizationProgressCallback(PyObject* progress_callback, const double percent_complete, const double seconds_elapsed, const double estimated_seconds_remaining, const long gcodes_processed, const long lines_processed);
+static bool ExecuteStabilizationProgressCallback(PyObject* progress_callback, const double percent_complete, const double seconds_elapsed, const double estimated_seconds_remaining, const int gcodes_processed, const int lines_processed);
 static bool ExecuteGetSnapshotPositionCallback(PyObject* py_get_snapshot_position_callback, double x_initial, double y_initial, double* x_result, double* y_result);
 #endif
